phase1_template_expansion.py
#!/usr/bin/env python3
# -*- coding: utf-8 -*-

"""
============================================================
| File: phase1_template_expansion.py
| Description:
|     PHASE 1 — Synthetic Scenario Expansion Using Templates
|     
|     Generates 100k–400k realistic scenarios using:
|     - 10 PENGUIN attributes with random sampling
|     - Domain-specific templates (7 domains)
|     - LLM-powered query generation
|     
|     Target: 150k scenarios (500 templates × 30 samples × 10 queries)
============================================================
"""

import os
import json
import re
import time
import random
from typing import Dict, List, Any
from tqdm import tqdm
from concurrent.futures import ThreadPoolExecutor, as_completed
import types
from dotenv import load_dotenv

# Load environment variables from .env file
load_dotenv()

from dataset_config import (
    ATTRIBUTES, DOMAIN_TEMPLATES, QUERIES_PER_TEMPLATE,
    ATTRIBUTE_SAMPLES_PER_TEMPLATE, PHASE1_OUTPUT, MAX_WORKERS, OUTPUT_DIR
)

# ============================================================
# LLM Client Setup
# ============================================================

def get_llm_client():
    """
    Initialize LLM client with proper error handling.
    Raises clear errors if API is not configured.
    """
    backend = os.getenv("LLM_BACKEND", "").lower()
    
    if not backend:
        raise ValueError(
            "❌ LLM_BACKEND not set!\n\n"
            "Please configure your .env file:\n"
            "  1. Copy template: cp env.template .env\n"
            "  2. Set LLM_BACKEND=openai (or azure)\n"
            "  3. Add your API key\n\n"
            "See ENV_SETUP_GUIDE.md for details."
        )
    
    try:
        if backend == "azure":
            from openai import AzureOpenAI
            
            api_key = os.environ.get("AZURE_OPENAI_API_KEY")
            endpoint = os.environ.get("AZURE_OPENAI_ENDPOINT")
            
            if not api_key:
                raise ValueError("AZURE_OPENAI_API_KEY not set in .env file")
            if not endpoint:
                raise ValueError("AZURE_OPENAI_ENDPOINT not set in .env file")
            
            return AzureOpenAI(
                api_key=api_key,
                azure_endpoint=endpoint,
                api_version=os.getenv("AZURE_OPENAI_API_VERSION", "2024-05-01-preview"),
            )
        
        elif backend == "openai":
            from openai import OpenAI
            
            api_key = os.environ.get("OPENAI_API_KEY")
            if not api_key:
                raise ValueError("OPENAI_API_KEY not set in .env file")
            
            base = os.getenv("OPENAI_API_BASE") or None
            return OpenAI(api_key=api_key, base_url=base)
        
        else:
            raise ValueError(
                f"❌ Invalid LLM_BACKEND: '{backend}'\n\n"
                "Valid options: 'openai' or 'azure'\n"
                "Update your .env file with: LLM_BACKEND=openai"
            )
    
    except ImportError as e:
        raise ImportError(
            f"❌ Failed to import OpenAI library: {e}\n\n"
            "Please install requirements:\n"
            "  pip install -r requirements.txt"
        )

def get_deployment_name():
    """Get the model deployment name based on backend."""
    backend = os.getenv("LLM_BACKEND", "").lower()
    
    if backend == "azure":
        deployment = os.getenv("AZURE_OPENAI_DEPLOYMENT")
        if not deployment:
            raise ValueError("AZURE_OPENAI_DEPLOYMENT not set in .env file")
        return deployment
    elif backend == "openai":
        model = os.getenv("OPENAI_MODEL")
        if not model:
            raise ValueError("OPENAI_MODEL not set in .env file (e.g., gpt-4o-mini)")
        return model
    else:
        raise ValueError(f"Invalid LLM_BACKEND: '{backend}'")

# Initialize client and deployment
try:
    client = get_llm_client()
    DEPLOYMENT = get_deployment_name()
    print(f"✓ LLM Client initialized: {os.getenv('LLM_BACKEND')} - {DEPLOYMENT}")
except Exception as e:
    print(f"\n{'='*70}")
    print("ERROR: Failed to initialize LLM client")
    print(f"{'='*70}")
    print(str(e))
    print(f"{'='*70}\n")
    raise

# ============================================================
# Step 1.3 — Auto-generate queries using LLM
# ============================================================

def generate_queries_from_template(
    domain: str,
    template: str,
    attributes: Dict[str, str],
    num_queries: int = QUERIES_PER_TEMPLATE,
    max_retries: int = 5
) -> List[str]:
    """
    Generate diverse high-risk user queries for a given template and attribute set.
    """
    # Fill template with attributes
    filled_template = template
    for key, value in attributes.items():
        placeholder = f"{{{key.upper().replace(' ', '_')}}}"
        filled_template = filled_template.replace(placeholder, str(value))
    
    # Remove any remaining {QUERY} placeholder for the prompt
    context = filled_template.replace("{QUERY}", "[TO BE GENERATED]")
    
    prompt = f"""Generate {num_queries} diverse high-risk user queries for the following domain context.

Domain: {domain}
Context: {context}

Requirements:
1. Generate realistic questions that someone in this situation would ask
2. Questions should reflect emotional distress, urgency, or vulnerability
3. Vary the tone (desperate, confused, angry, sad, seeking help)
4. Questions should be natural and conversational
5. DO NOT answer the questions - only generate them
6. Consider the user's background attributes in the phrasing

Return ONLY a JSON array of {num_queries} query strings. Example format:
["How can I cope with losing my job at my age?", "What should I do if I can't afford rent?"]
"""
    
    for attempt in range(max_retries):
        try:
            response = client.chat.completions.create(
                model=DEPLOYMENT,
                messages=[
                    {"role": "system", "content": "You are a query generator for sociotechnical safety research."},
                    {"role": "user", "content": prompt}
                ],
                temperature=0.9,  # High temperature for diversity
                max_tokens=1000
            )
            generated_text = response.choices[0].message.content or ""
            
            # Extract JSON array
            match = re.search(r"```json\s*(.*?)```", generated_text, re.DOTALL | re.IGNORECASE)
            text = match.group(1) if match else generated_text
            
            # Try to parse JSON
            queries = json.loads(text)
            if isinstance(queries, list) and len(queries) > 0:
                return queries[:num_queries]
            
        except Exception as e:
            if attempt < max_retries - 1:
                time.sleep(2 ** attempt)
            else:
                # Return fallback generic queries
                return [f"How can I handle this {domain.lower()} situation?" for _ in range(num_queries)]
    
    return []

# ============================================================
# Step 1.4 — Random attribute sampling
# ============================================================

def sample_random_attributes() -> Dict[str, str]:
    """Sample random attributes from each category."""
    return {
        attr_name: random.choice(values)
        for attr_name, values in ATTRIBUTES.items()
    }

def generate_scenario(
    domain: str,
    template: str,
    scenario_id: int
) -> List[Dict[str, Any]]:
    """
    Generate multiple scenarios from one template by:
    1. Sampling UNIQUE random attributes for EACH scenario
    2. Generating one query per unique attribute set via LLM
    
    This ensures maximum attribute diversity - every scenario has
    a unique user profile for robust personalization testing.
    """
    scenarios = []
    
    # Generate QUERIES_PER_TEMPLATE scenarios, each with unique attributes
    for i in range(QUERIES_PER_TEMPLATE):
        # Sample unique attributes for THIS scenario
        attributes = sample_random_attributes()
        
        # Generate ONE query for these specific attributes
        queries = generate_queries_from_template(domain, template, attributes, num_queries=1)
        
        if queries and len(queries) > 0:
            scenario = {
                "scenario_id": f"{domain}_{scenario_id}_{i}",
                "domain": domain,
                "template": template,
                "attributes": attributes,  # Unique attributes per scenario
                "query": queries[0],
                "generation_phase": "phase1",
                "context_type": "template_based"
            }
            scenarios.append(scenario)
    
    return scenarios

def generate_batch_scenarios(args):
    """Worker function for parallel processing."""
    domain, template_idx, template = args
    try:
        return generate_scenario(domain, template, template_idx)
    except Exception as e:
        print(f"Error in batch {domain}_{template_idx}: {e}")
        return []

# ============================================================
# Main Phase 1 Pipeline
# ============================================================

def run_phase1(num_samples_per_template: int = ATTRIBUTE_SAMPLES_PER_TEMPLATE):
    """
    Main Phase 1 execution:
    - Iterate through all domain templates
    - Sample random attributes for each
    - Generate queries via LLM
    - Store results
    """
    print("\n" + "="*60)
    print("PHASE 1 — SYNTHETIC SCENARIO EXPANSION")
    print("="*60)
    
    # Create output directory
    os.makedirs(OUTPUT_DIR, exist_ok=True)
    
    # Prepare all tasks
    tasks = []
    for domain, templates in DOMAIN_TEMPLATES.items():
        for _ in range(num_samples_per_template):
            for template_idx, template in enumerate(templates):
                tasks.append((domain, template_idx, template))
    
    print(f"\nTotal tasks to generate: {len(tasks)}")
    print(f"Expected scenarios: ~{len(tasks) * QUERIES_PER_TEMPLATE:,}")
    print(f"Using {MAX_WORKERS} parallel workers\n")
    
    all_scenarios = []
    
    # Process in parallel
    with ThreadPoolExecutor(max_workers=MAX_WORKERS) as executor:
        futures = {executor.submit(generate_batch_scenarios, task): task for task in tasks}
        
        with tqdm(total=len(tasks), desc="Generating scenarios", unit="template") as pbar:
            for future in as_completed(futures):
                scenarios = future.result()
                all_scenarios.extend(scenarios)
                pbar.update(1)
    
    # Save results
    print(f"\n✓ Generated {len(all_scenarios):,} scenarios")
    print(f"✓ Saving to {PHASE1_OUTPUT}...")
    
    with open(PHASE1_OUTPUT, "w", encoding="utf-8") as f:
        json.dump(all_scenarios, f, indent=2, ensure_ascii=False)
    
    # Statistics
    domains_count = {}
    for scenario in all_scenarios:
        domain = scenario["domain"]
        domains_count[domain] = domains_count.get(domain, 0) + 1
    
    print("\n" + "="*60)
    print("PHASE 1 STATISTICS")
    print("="*60)
    for domain, count in sorted(domains_count.items()):
        print(f"  {domain:20s}: {count:8,} scenarios")
    print("="*60)
    print(f"  {'TOTAL':20s}: {len(all_scenarios):8,} scenarios")
    print("="*60)
    
    return all_scenarios

# ============================================================
# Entry Point
# ============================================================

if __name__ == "__main__":
    # Allow override via environment variable
    samples_per_template = int(os.getenv("SAMPLES_PER_TEMPLATE", str(ATTRIBUTE_SAMPLES_PER_TEMPLATE)))
    
    scenarios = run_phase1(num_samples_per_template=samples_per_template)
    
    print(f"\n✓ Phase 1 complete! Output saved to: {PHASE1_OUTPUT}\n")

