requirements.txt

openai>=1.40.0
python-dotenv>=1.0.1
numpy>=1.26.0
pandas>=2.2.0
tqdm>=4.66.0

# Optional: For Anthropic Claude support
# anthropic>=0.39.0
